competible with init part
