--- irrationalAgents/API/unity/codes
@@ -1,24 +1,0 @@
-ui.tick
-server.tick
-
-player.getInfo
-player.getInfo.response
-
-npc.getList
-npc.getList.response
-
-npc.getInfo
-npc.getInfo.response
-
-npc.navigate
-
-map.getTownData
-map.getTownData.response
-
-map.getSceneMetadata
-map.getSceneMetadata.response
-
-config.getBlockData
-config.getBlockData.response
-
-chat.updateNPC
